requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Brotli==1.1.0
click==8.1.7
colorama==0.4.6
colorgram.py==1.2.0
dj-database-url==2.2.0
Django==4.2.15
django-cors-headers==4.4.0
django-grappelli==4.0.1
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
h11==0.14.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
waitress==3.0.0
whitenoise==6.7.0
